+++ py/desiutil/io.py
@@ -106,3 +106,114 @@ def yamlify(obj, debug=False):
     if debug:
         print(type(obj))
     return obj
+
+def _dtype_size(dtype):
+    '''
+    Parse dtype like '<Un' into int(n)
+    Note that this is different from dtype.itemsize, which is number of bytes
+    '''
+    i = dtype.str.find(dtype.kind)
+    return int(dtype.str[i+1:])
+
+def _pick_encoding(table, encoding):
+    '''
+    pick which encoding to use; giving warning if options are in conflict
+
+    Args:
+        table : astropy Table object
+        encoding : (str) encoding to use; None to use table.meta['ENCODING']
+
+    Note: `encoding` trumps `table.meta['ENCODING']`
+    '''
+    if encoding is None:
+        if 'ENCODING' in table.meta:
+            encoding = table.meta['ENCODING']
+        else:
+            raise UnicodeError('No encoding given as argument or in table metadata')
+    elif 'ENCODING' in table.meta and table.meta['ENCODING'] != encoding:
+        import warnings
+        message = """\
+data.metadata['ENCODING']=='{}' does not match option '{}';
+use encoding=None to use data.metadata['ENCODING'] instead""".format(\
+            table.meta['ENCODING'], encoding)
+        warnings.warn(message)
+
+    return encoding
+
+def encode_table(data, encoding='ascii'):
+    '''
+    Encode unicode strings in a table into bytes using numpy.char.encode
+
+    Args:
+        data : numpy structured array or astropy Table
+
+    Options:
+        encoding : encoding to use for converting unicode to bytes.
+            Default 'ascii' (FITS and HDF5 friendly), but if None,
+            use ENCODING from table metadata if available
+
+    Returns astropy Table with unicode columns converted to bytes
+
+    Raises:
+        UnicodeEncodeError if any input strings cannot be encoded using
+            the specified encoding
+        UnicodeError if no encoding is given as argument or in table metadata
+
+    Note: `encoding` option overides data.meta['ENCODING'];
+        use encoding=None to use data.meta['ENCODING'] instead
+    '''
+    from astropy.table import Table
+    import numpy as np
+
+    try:
+        table = Table(data, copy=False)
+    except ValueError:  #- https://github.com/astropy/astropy/issues/5298
+        table = Table(data, copy=True)
+
+    encoding = _pick_encoding(table, encoding)
+
+    for col in table.colnames:
+        dtype = table[col].dtype
+        if dtype.kind == 'U':
+            Sn = 'S{}'.format(_dtype_size(dtype))
+            table.replace_column(col, np.char.encode(table[col], encoding=encoding).astype(Sn))
+
+    table.meta['ENCODING'] = encoding
+    return table
+
+def decode_table(data, encoding='ascii', native=True):
+    '''
+    Decode byte strings in a table into unicode strings
+
+    Args:
+        data : numpy structured array or astropy Table
+
+    Options:
+        encoding : encoding to use for converting bytes into unicode;
+            default 'ascii'; if None, try ENCODING keyword in data instead
+        native : if True (default), only decode if native str type is unicode
+            (i.e. python3 but not python2)
+
+    Note: `encoding` option overides data.meta['ENCODING'];
+        use encoding=None to use data.meta['ENCODING'] instead
+    '''
+    from astropy.table import Table
+    import numpy as np
+    try:
+        table = Table(data, copy=False)
+    except ValueError:  #- https://github.com/astropy/astropy/issues/5298
+        table = Table(data, copy=True)
+
+    #- Check if native str type is bytes
+    if native and np.str_('a').dtype.kind == 'S':
+        return table
+
+    encoding = _pick_encoding(table, encoding)
+    for col in table.colnames:
+        dtype = table[col].dtype
+        if dtype.kind == 'S':
+            Un = 'U{}'.format(_dtype_size(dtype))
+            table.replace_column(col, np.char.decode(table[col], encoding=encoding).astype(Un))
+
+    table.meta['ENCODING'] = encoding
+    return table